+++ webui/react/src/pages/F_ExpList/glide-table/custom-cells/avatar.tsx
