+++ webui/react/src/pages/F_ExpList/glide-table/custom-cells/sparkline.tsx
